Remove `ClojureLambdaBase`.
This base class was previously needed to ensure that class loading
worked properly. Although I have not been able to find any offical
docs/changelog to support this, it now seems like the AWS implementation
has changed and the class loader is set as expected. This means we no
longer need this hack.